# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Auth API",
    "version": "1.0.0",
    "description": "API documentation with JWT authentication"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "User Signup",
        "description": "Register a new user with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "testuser@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password123!"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Login",
        "description": "Authenticate user and return JWT access and refresh tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "testuser@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password123!"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success - returns access_token and refresh_token"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "User Logout",
        "description": "Log out current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public Info",
        "description": "Returns public welcome message",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get User Profile",
        "description": "Returns user profile details for authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}